+++ autorest/azure/environments.go
@@ -71,7 +71,7 @@ type Environment struct {
 	ContainerRegistryDNSSuffix        string             `json:"containerRegistryDNSSuffix"`
 	CosmosDBDNSSuffix                 string             `json:"cosmosDBDNSSuffix"`
 	TokenAudience                     string             `json:"tokenAudience"`
-	ApiManagementHostManagementSuffix string             `json:"apiManagementHostManagementSuffix"`
+	APIManagementHostManagementSuffix string             `json:"apiManagementHostManagementSuffix"`
 	ResourceIdentifiers               ResourceIdentifier `json:"resourceIdentifiers"`
 }
 
@@ -99,7 +99,7 @@ var (
 		ContainerRegistryDNSSuffix:        "azurecr.io",
 		CosmosDBDNSSuffix:                 "documents.azure.com",
 		TokenAudience:                     "https://management.azure.com/",
-		ApiManagementHostManagementSuffix: "azure-api.net",
+		APIManagementHostManagementSuffix: "azure-api.net",
 		ResourceIdentifiers: ResourceIdentifier{
 			Graph:               "https://graph.windows.net/",
 			KeyVault:            "https://vault.azure.net",
@@ -133,7 +133,7 @@ var (
 		ContainerRegistryDNSSuffix:        "azurecr.us",
 		CosmosDBDNSSuffix:                 "documents.azure.us",
 		TokenAudience:                     "https://management.usgovcloudapi.net/",
-		ApiManagementHostManagementSuffix: "azure-api.us",
+		APIManagementHostManagementSuffix: "azure-api.us",
 		ResourceIdentifiers: ResourceIdentifier{
 			Graph:               "https://graph.windows.net/",
 			KeyVault:            "https://vault.usgovcloudapi.net",
@@ -167,7 +167,7 @@ var (
 		ContainerRegistryDNSSuffix:        "azurecr.cn",
 		CosmosDBDNSSuffix:                 "documents.azure.cn",
 		TokenAudience:                     "https://management.chinacloudapi.cn/",
-		ApiManagementHostManagementSuffix: "azure-api.cn",
+		APIManagementHostManagementSuffix: "azure-api.cn",
 		ResourceIdentifiers: ResourceIdentifier{
 			Graph:               "https://graph.chinacloudapi.cn/",
 			KeyVault:            "https://vault.azure.cn",
@@ -201,7 +201,7 @@ var (
 		ContainerRegistryDNSSuffix:        NotAvailable,
 		CosmosDBDNSSuffix:                 "documents.microsoftazure.de",
 		TokenAudience:                     "https://management.microsoftazure.de/",
-		ApiManagementHostManagementSuffix: NotAvailable,
+		APIManagementHostManagementSuffix: NotAvailable,
 		ResourceIdentifiers: ResourceIdentifier{
 			Graph:               "https://graph.cloudapi.de/",
 			KeyVault:            "https://vault.microsoftazure.de",